Add an optional argument to useIntervalEffect to synchronise updates
Sadly not got time to do the documentation and tests for this in the next couple of weeks at least - but if anyone wants to add those bits to get this to a mergeable point then go for it!

This adds an optional synchronisation to useIntervalEffect - in other words, anything that uses the option and has the same update time will be called in the same update cycle. For instance allowing for multiple clocks on screen that all update in the same tick etc.

It uses two Maps, one simply stores the `setInterval` result so it can be removed when unused, and the second stores an array of callbacks. Whenever a callback is added or removed the array is mutated (rather than replaced - GC reasons mainly), and the callbacks themselves are wrapped in try/catch so they don't break anything. The dependencies list adds the flag coerced into a boolean - don't care what it actually is, but `undefined == false` and hooks use `===` so it saves processing on that edge case of `undefined` -> `false`.