fix(useStorageValue): Set initializeWithValue to false by default
This option was previously true by default, which is in conflict with this library's goal of
supporting SSR out of the box. Also the documentation and the tests seem to presume that
initializeWithValue is false by default. This is not a breaking change since due to a bug (#1002)
initializeWithValue was effectively false by default already.

re #1002