fix(useMeasure): Only expose width and height
Only the width and height of the observed elements content rect are now exposed because the other
properties were not considered relevant for this hook. In addition they introduced potential
footguns. See #954 for more.

BREAKING CHANGE: Only width and height are now exposed from the element instead of the whole DOMRect
#951